==> analysis/main/5-1-prop-bot-reverts.ipynb <==
# Section 5.1: proportion of all bot edits to articles that are bot-bot reverts

This is a data analysis script for an analysis presented in section 5.1, which you can run based entirely off the files in this GitHub repository. It loads `datasets/montly_bot_edits/[language]wiki_20170427.tsv` and `datasets_monthly_bot_reverts/[language]wiki_20170420.tsv`.

import pandas as pd
import seaborn as sns
import mwapi
import numpy as np
import glob
%matplotlib inline

# Import datasets
Using a dictionary of pandas dataframes, with the key as the language. A better way would be to have a tidy dataframe.

!ls -lah ../../datasets/monthly_bot_edits/

!ls -lah ../../datasets/monthly_bot_reverts/

Monthly bot edit counts are in the format: month (YYYYMM), page namespace, and total number of bot edits in that language's namespace that month (n).

!head ../../datasets/monthly_bot_edits/enwiki_20170427.tsv

Monthly bot revert counts are a bit more complicated and in a slightly different format: 
- month (YYYYMM01)
- page namespace
- number of total reverts by all editors (reverts)
- number of reverts by bot accounts (bot_reverts)
- number of edits by bots that were reverted (bot_reverteds)
- number of reverts by bots of edits made by bots (bot2bot_reverts)

!head ../../datasets/monthly_bot_reverts/enwiki_20170420.tsv

### Processing

We're going to do this in a pretty messy and so-not-best-practice way (which would be a single tidy dataframe with nice hierarchical indexes) by having two dictionaries of dataframes. 

df_edits_dict = {}
for filename in glob.glob("../../datasets/monthly_bot_edits/??wiki_2017042?.tsv"):
    lang_code = filename[33:35]
    df_edits_dict[lang_code] = pd.read_csv(filename, sep="\t")
    df_edits_dict[lang_code] = df_edits_dict[lang_code].drop_duplicates()

for lang, lang_df in df_edits_dict.items():
    print(lang, len(lang_df))

df_rev_dict = {}
for filename in glob.glob("../../datasets/monthly_bot_reverts/??wiki_2017042?.tsv"):
    lang_code = filename[35:37]
    df_rev_dict[lang_code] = pd.read_csv(filename, sep="\t")
    df_rev_dict[lang_code] = df_rev_dict[lang_code].drop_duplicates()

for lang, lang_df in df_rev_dict.items():
    print(lang, len(lang_df))

langs = ["de", "en", "es", "fr", "ja", "pt", "zh"]

### Preview the dataframes in the dictionaries

df_edits_dict['en'][0:5]

df_rev_dict['en'][0:5]

# Clean and combine the two datasets

## Convert dates

Remember that they used different formats for representing months? Gotta fix that.

def truncate_my(s):
    """
    Truncate YYYYMMDD format to YYYYMM. For use with df.apply()
    """
    s = str(s)
    return int(s[0:6])

Test function

truncate_my(20100101)

Yay!

### Apply the transformation

for lang in langs:
    df_edits_dict[lang] = df_edits_dict[lang].set_index('month')
    df_rev_dict[lang]['month_my'] = df_rev_dict[lang]['month'].apply(truncate_my)
    df_rev_dict[lang] = df_rev_dict[lang].set_index('month_my')

### Combine the datasets, looking at only articles / ns0

combi_ns0_dict = {}
combi_dict = {}
for lang in langs:
    print(lang)

    combi_ns0_dict[lang] = pd.concat([df_rev_dict[lang].query("page_namespace == 0"), df_edits_dict[lang].query("page_namespace == 0")], axis=1, join='outer')
    combi_ns0_dict[lang]['bot_edits'] = combi_ns0_dict[lang]['n']
    combi_ns0_dict[lang]['prop_bot2bot_rv'] = combi_ns0_dict[lang]['bot2bot_reverts']/combi_ns0_dict[lang]['bot_edits']

### Preview the combined dataframe
FYI, all things Wikipedia database related (especially bots) are generally way less consistent before 2004.

combi_ns0_dict['en'][29:39]

## The results: proportion of bot-bot reverts out of all bot edits, articles/ns0 only

sum_dict = {}
for lang in langs:
    #print(lang)
    sum_dict[lang] = combi_ns0_dict[lang][['bot_edits','bot2bot_reverts']].sum()
    print(lang, "ns0 proportion:", (sum_dict[lang]['bot2bot_reverts']/sum_dict[lang]['bot_edits']*100).round(4), "%")
    print(sum_dict[lang])
    print("")
